move ReducerFn type
